Fix baseline for Zeek 4.0.0+ output
The change is related to https://github.com/zeek/zeek/pull/1160.  HTTP
DPD signatures no longer require matching both sides of a connection, so
the `json-rpc-request-http` signature which relies on
`http-request-body` is allowed to trigger on some connections where it
previously did not.  Ultimately, the test case shows that the
scripts/signature still catch the same miner and the same pool-servers,
but on earlier connections than before, which generally may still be ok
(but could mean more false positive detections of pool servers).